requirements.txt
pandas
scikit-learn
streamlit
joblib
python-dotenv
watchdog
